requirements.txt
customtkinter
Aspose.Email-for-Python-via-NET
packaging
